refactor (refs T28977): update EDT to 0.12.13
The refactored warnings are currently no longer triggered in the
application, thus exceptions can be used instead to prevent
errorneous usage in the future.

Provide MessageFormatter where needed.

Provide API request handling with URL params.

Use validators in EDT for API request handling:

Validating the URL parameters `fields` and `includes` is now done by
using superior logic provided by classes in EDT. Some checks were set
to be more strict (exceptions instead of logging), because there were
no occurrences of the logging messages in the last 3 months on any
server.

In previous EDT version the validation of property paths was incomplete.
The last path segment was always left out from the validation. E.g. when
sorting via `procedure.name` it was not checked if `name` is allowed to
be used to sort. Likewise, when including the procedure of a requested
statement via the path `procedure` it was not checked if the `procedure`
property is set as `readable` or if the referenced `Procedure` resource
type is set as available.

The problem in EDT is fixed now, but because there are requests in dplan
accessing properties that should not be available, a custom
`PropertyPathProcessor` was added, which allows the old
(invalid) access, while logging the violation. Thus, after the requests
have been fixed the `DplanPropertyPathProcessor` and
`DplanPropertyPathProcessorFactory` classes can be removed.

In case of the `ApiController` the new classes were initialized instead
of being injected, as injection would have required to adjust multiple
inheriting controller classes.