feature: automatically remove invalid assignees on procedure edit
Context:

When editing a procedure you may change things like the authorized
users. However till now even though a user was no longer authorized
for a procedure afterwards, they were still set as assignees on
previously claimed statements and segments. In this PR the
ProcedureEditedEvent is used to check for invalid statement/segment
claims and unassign them.

Solution:

To determine which assignees are no longer valid the logic inside
the isOwningProcedure method needed to be used. However, as the
implementation of isOwningProcedure was inadequate it was
duplicated into getOwnsProcedureCondition. Code duplication was
sufficiently avoided by moving almost all shared code into a new
OwnsProcedureConditionFactory class.

Bugfix:

During the implementation it became clear that in a previous
refactoring the consideration of the current customer was
accidentally removed from the ownsProcdure evaluation, which
affects permission checks. This PR re-adds the check.