Fix: Remove lila messages
Some fields from Statement.js arent existing any more in the Resourcetypes, so they arent needed in the projects.
Remove them from the fields in Statement.js to fix the lila messages from the `Abwägungstabelle`.
Some fields are just available, if selected permissions are active.
PriorityArea and Mupicality are available if the currentuser has the permission of area_admin_assessmenttable in `Abwägungstabelle`.
So write an extra condition to check the permission in their resourcetypes.
In Statement.js write a condition that checks for the other available fields, if the selected permissions are active.