Fix: Remove the buttons in the header and make the buttons on buttom sticky
The same sticky buttons in the header are shown in the buttom.
The buttom bottons werent sticky. It was confusing the user.
Normally the user is expecting these buttons at the bottom of the page.
So remove the buttons in the header.
The buttons in the bottom should be sticky so that the user can see them if he is in the top of page.
Little change in the scss file for better understanding for the usage.