fix(refs T28957): reset orga when appropriate
Before, the user data of all DpUserFormField instances were reset
whenever users clicked the "reset" button on the form to create
a new user, which led to all organisation dropdowns being blanked.

This is fixed by passing the `resetFromRoot` prop to DpUserFOrmFields.

To reset the organisation dropdown whenever a user list item
is closed (= to fix the actual bug), the `localUser` data field
is watched from within userFormFields.

Lastly, the reset button on the organisation list item is renamed from
"close" to its default value (reset), because this is what actually happens.