feat(refs T28813): add LICENSE file with 3rd party notes
The first line of the LICENSE file, "All rights reserved. Usage
only allowed with distinct Agreement.", is a non-permissive
phrase that will be replaced when demosplan is released
under the appropriate open source license.

Following are license notices required by 3rd party software
used by demosplan not listed in /licenses directory. These follow
a commonly used scheme (at least I found it commonly used
when researching it), if applicable:

   - path to the 3rd party code inside the source
   - name of the library (if there is one single name existing)
   - short form of license the library is released under
   - full license text (omitted when multiple licenses apply)

The notices will eventually be merged with the generated notices
in the /licenses directory later on, in a manner that gets rid of
the redundant license texts.

As the rateIt library notice is included here, it can be dropped
from the vendorlistUpdateCommand.

The mistakenly placed demosplan license headers within the
Detabinator and SideNav files is removed, as those files
already contain the correct 3rd party license header.