fix: Involve a try-catch if the value of `$filterHash` gets no update
Sometimes you get an Error, if you would like to click into a procedure and go to the overview page.
This happens because the value of `$filterHash` doesnt get updated,
because the `hashList`, which you need for `getSession` does not update.
To solve this, write a try-catch in the condition and if the hashList does not update,
then it is able to continue with the default value of `$filterHash`.