fix(refs T28625): improve message consistency
Before, the confirm message that appeared before sending
the mails was not in sync with the success messages appearing
after the message was sent. Also, there was some redundancy in
the success messages. This is solved by:

- explaining what will happen more explicitly within the
  `check.mail.result` messages (note that for institution
  statements voting is not possible)
- diversify and compose the confirmation message: the first
  part is now chosen based on who the email is sent to, and
  the second part is statically added as a new message. The
  `other` case is actually not needed as it resolves to
  `citizen_only`, however is also set as a fallback because
  it is  the most neutral text

$successMessageTranslationParams may be verbose but as it needs to be declared
upfront (without the context of its usage) i opted to name it that way.