refactor: Remove relative import references
As we're moving our JS to `client/js` path lookups change.
Since we don't want to move everything at once, the relative
paths used for some imports in bundles need to be made absolute
and refer to their bundle alias.

This will of course over time be refactored again to all
reference starting from @DpJs but for now this seems like
the best solution.