feat: Support loading bundles from client/js
Loading bundles from `client/js` is a bit tricky since not all
code loaded by the bundles may be there already.

The alias config needs to refer to the old bundle locations
as it is used to lookup components still referring to @DemosPlan<bundle>Bundle.

The entrypoints config however needs to add the new bundle locations
to add them to the compiled entrypoints.