fix (refs T29701): disables default ES listeners for segments
We manage the ES updates based on Doctrine lifecycle events with our own listener. Our custom listener creates SearchIndexTasks in the database that will be used to update the ES on a Symfony ResponseEvent. Therefore we can't allow the default ES listeners to be active as they will update the ES instantly which leads to the ES being out of sync with Doctrine when Doctrine changes are not committed (because of an exception, error or rollback).