fix: Set the value to hasPriorityAreas in obj in DpTable, push the fields in the includes array due to his permission and add their name in the fields of statement, set the condition correct for hasPriority when currentUser has not the permission `area-admin-assessmenttable`, remove the wrong bracket part in assessment_table_view.html.twig
In DpTable it is necessary to set the data procedureStatementPriorityArea from the prop from assessment_table_view.html.twig as value for hasPriorityArea,
so it can be used in store Statement.js. The include fields in statement store should only be included if the correct permission is active.
Due to priorityAreas the data from the getStatementAction request should be true to get included.
In assessment_table_view.html.twig because of the wrong closing bracket the boolean get true.
In PriorityAreaResourceType.php the condition for hasPermission('area_admin_assessmenttable') needs to be changed that if the user has not this
permission it goes to the following checks.