feat: adds ExecuteAddonExtension.php
The twig extension formerly known as 'ExecutePluginExtension' is reintroduced here. It was removed with most of the old PluginBundle but as it turns out we need it to allow for some twig extension capabilities. To get back to a working state asap, I propose to add it back in and reevaluate it at a later time.