refactor: reduce `ProcedurePerson` initialization complexity
To reduce code duplication the non-API request
handling re-used the existing
`createPersonAndAddToStatementWithResourceType`
method, which already existed for API purposes.
This however comes with too much unneeded complexity.
Thus, the call was replaced with the logic actually
needed. This may also fix a tooling concern regarding
the potential propagation of potentially malicious
content, which does not seem true, as the request
values are validated via Symfony validation
annotations in the entity after being set into the
entity but before finishing the database transaction.