fix: do not exclude all dep sourcemaps
In Chrome devtools, there was a flood of messages with
"DevTools failed to load SourceMap" warnings which made
spotting the real logs harder. This seems to be an issue
of sentry/browser and a few other libs that reference
a sourcemap by url within their esm files. The workaround is
to not exclude those modules from source-map-loader.

See https://github.com/getsentry/sentry-javascript/issues/2566