feature (ADO 1740): restructure permission initialization
Permissions will be now fetched by the core and evaluated
there. `PermissionEvaluator` instances beside the one
provided by the core were removed.

Addons are now allowed to evaluate any permission
(via the core `PermissionEvaluator`), even if it was
defined by an addon the evaluating one did not define
as dependency.