fix(refs T30430): use GitHub hash to reference demosplan-ui 0.0.6
the v0.0.6 release of demosplan-ui did not contain the actual code
of that release, as we confused the "prepublish" with the appropriate
"prepare" npm lifecycle hook. Until the next release fixes that, this
commit requires an interim branch in which demosplan-ui moves webpack,
webpack-cli and style-dictionary within its dependencies (leading to a
temporarily bloated "yarn.lock" but things will settle as the 0.0.7
npm version of demosplan-ui comes along).

See https://docs.npmjs.com/cli/v9/using-npm/scripts