fix(refs T30831): Check permissions for Tags
Tags are not available in every project.
Therefore we have to check the permissions before adding
tags to the statement request.

Since we get more and more checks, I refactored the logic a bit, that we have only the if-block, where the data can be added.

BUT: We should think about how to handle that in a more generic way.