feat (refs T30460): adds NonAuthorizedCaseworkerRemover
It can happen that a user was assigned as caseworker for a number of statements in a procedure, but was later removed as an authorized user for that procedure. When that happens, we need to check and remove mentions of this user from all statements where the user is still the caseworker.