feat: adds 2 mandatory methods
The interfaces now enforce 2 more methods that were missing. This should take care of that. Further adjustments to see if we could replace other things with the new methods is outside the scope of this PR and should be done at a later time or when we touch the filesystem the next time.