fix (refs T30990): Escape any special chars for output
In most cases we store texts with already escaped special chars in the database to avoid any later problems. But in some rare cases and for reasons I don't, some special chars make their way into the database. Without the here added line of code (which is present in most if not all other docx exports) those will break the docx generation as that is based on a xml file.