chore: Remove unused code
The QueryStorageLoader was previously used to handle the loading
of the different stored queries which we use to handle complex
query strings in the assessment table and the newer statement/segment lists.

However, that loading has long since been moved into
\demosplan\DemosPlanCoreBundle\Doctrine\Type\StoredQueryType
and is thus handled transparently by Doctrine when
entities which have stored queries (e.g. Procedure) are requested/stored.