Improve addon contract wiring (#683)
* doc: Explain autoloading rationale

* refactor: Move composer related constant

* chore: Reconfigure addon directory ignoring

* refactor: Rewrite addon loading

Installing and loading Addons in demosplan is a tedious multistep process.
Some actions happen on the file system, some in Composer, some in Symfony.
Of the actions in Symfony, some need to occur during the container compile
step, some always and some after.

Previously, all of this was more or less mangled into the AddonRegistry
and the AddonInstallFromZipCommand. This commit is a step towards cleaning
things up and making the process a little more understandable and in
consequence hopefully also more maintainable.

As of now, to install an addon zip bundle (installation via pure
composer package is still not fully implemented), the AddonInstallFromZipCommand
has the logic for

- unarchiving the zip to the addons cache (which, as a reminder, is only
  responsible for storing zipped addons and will not be utilized for
  composer-only addons)
- handling the composer "installation" aka the local package registration
  and dependency resolving of the addon
- leaving the system bootable with prepared autoloading for addons

it then passes on to the Registrator which

- takes care of ensuring the addon is valid (not very comprehensive
  yet but that shall be extended upon in due time)
- loads and stores the addon's manifest into the addons.yaml

next, the AddonBuildFrontendCommand takes over and

- leaves again if there is no frontend configuration necessary, or

- installs the required frontend dependencies with Yarn
- runs the production build for the frontend assets assuming the
  addon was packaged using @demos-europe/demosplan-addon

Furthermore, as previously the autoloading, the bundle loading
was extracted into a standalone system.

The AddonRegistry is now an actually runtime usable registry
of all availble addons and configured during container compilation
to receive the AddonInfo instances which replace the AddonInfoProvider
system.

* fix: Ensure we always use the correct application

Since we're extending on Symfony's ConsoleApplication
with a few useful tidbits for our use cases, we should
enforce that we're actually receiving that extended object
instead of the OG one in console commands.

* feat: Make permission initializers available

* refactor: Replace iterators with simpler injection

We can use the AddonRegistry now to request the Addon's
PermissionInitializers.

* wip: begin missing autowiring

* bug(refs T30480): commit current WIP state

- basic implementation of the rpc route which will return the vue component as script etc
- basic implementation of the addon script injection
- removal of $addonInfoProviders of config.yml (for dev purposes)

* fix: makes addon asset loader usable

This just removes all the little mistakes in the conception to make it actually usable.

* fix: removes unused getAllAddons method

* fix: removes unused name parameter

as this is handled by the array_map already and would only lead to more nesting.

* fix: reverts some accidental merge mistakes

I had already fixed these before, but that was somehow reverted while merging main into this branch. I re-reverted them back to my solutions.

* chore: Add note about the danger of manual edits

The addons.yaml dictionary is (mostly) managed from
code and should eventually not be touched manually
at all except when the machine inevitably does
something it wasn't supposed to do.

* feat: Add some validation

Since we can't really make any permission checks at that point yet,
we should at least verify that the parameter we need to resolve
the requested scripts is in the request.

* refactor: Use constant

* chore: Fix typo

Co-authored-by: Velimir Müller <velimir@demos-deutschland.de>
Co-authored-by: AlexMenschoid <meyer-schlichtmann@demos-deutschland.de>