feat(refs T31218): Fix docx encoding
The default output encoding of PhpWord is en_US which leads to issues
in post processing as they don't have all the beautiful umlauts.

In the future, we might need to make this dependent on the application
or project locale, but for now, setting it to german should be enough.