feat: adds first prototype fixtures
This is an experimental first attempt to build alice fixtures to see if(how we can do it. Turns out, that our data model is not well suited for this atm and that we would need a php based "glue pass" to connect a bunch of separate entities created by fixtures to actually reach a usable state.

It might be a good step in the right direction to do that anyway, but we're not sure yet. Maybe we should start cleaning up our data model first. We'll have to see.