fix(refs T31258): improve display of statement count
Showing both numbers (statement and originalStatement count) by default
turned out to be not so understandable. To improve this, original statements
are now only shown within a tooltip that is triggered when hovering the
numbers. Hiding the original statements by default is the better option
as users are mostly interested in this metric.