fix (refs T31331): move UI-logic into twig
Originally the statement entity was converted into
an associative array in the backend and the field
`feedback` overridden with the corresponding
translation value before passing the array into the
twig file.

However, `getStatement` did no
longer return the array representation of the statement
but the statement object instead, which caused a bug.
Returning the object is fine, because
we want to handle objects in the backend and avoid
unnecessary conversions.

As it would be very bad practice to write a translation
value into the `feedback` property of the (now object)
statement (as
it may be persisted by accident after refactorings), the
approach was changed and the logic determining flags and
translation values was moved into the twig file, as this
seems to be a frontend-code responsibility anyway.