bug(refs T30632): Set layer source for printlayer
Creating the layersource was removed for hidden layers because of performance, but we also need the layer sources for hidden layers. Otherwise when updating the map fields right after loading the page (eg with drawing in the map) we get an error, because not all print layers have a source.