bug(refs: T31984): prevent unintentional deletion of editor text when saving metadata
This prevents the user from unintentionally deleting an unsaved text by synchronizing the local
statement in StatementMeta.vue (which also emits the local statement when saving only metadata)
with the statements from store. The editor automatically updates the state of statements in the
store when registering an input. This only occurs when a statement has not been segmented already.