fix(refs:T32028): sync assignee to work around local state
Inside StatementMeta, the statement obj persists its initial
state, which it also bubbles back up onSave. So, when loading
the view with unassigned statement, and then assigning it,
the local statement in StatementMeta still has the unassigned
state which is applied to the outer state then, leading to the
buggy behavior.

To work around this, we can either only pass the data into
StatementMeta which is actually needed by this component, or
we can fix it afterwards (this version) - which is a little
ugly but as most data is needed within StatementMeta, it is
still the better choice.