fix(refs:T32028): remove outer button
The claim component generates a link element.
To wrap that within a button is not html5 conform,
also the link is not really the correct element here.
To still have the text label besides the icon, a new
prop is used in DpClaim. Also, some flex classes
are used to ensure vertical alignment also for the
loading state.