fix(refs:T32028): set local state only after successful saving
the optimistic behavior (that was implemented before) is not
the standard behavior within demosPlan. also, as we show a loading
indicator on the claim button, the rest of the ui should be in sync
with that. the alternative would have been to also remove the
loading indicator, hence hiding the loading completely. however
as we do not use this pattern at all, why should we use it here?