fix(refs T32028): Fix unclaiming of statement (#1068)
Fixing the bug
Inside StatementMeta, the statement obj persists its initial state, which it also bubbles back up onSave. So, when loading the view with unassigned statement, and then assigning it, the local statement in StatementMeta still has the unassigned state which is applied to the outer state then, leading to the buggy behavior.

To work around this, we can either only pass the data into StatementMeta which is actually needed by this component, or we can fix it afterwards (this version) - which is a little ugly but as most data is needed within StatementMeta, it is still the better choice.

Button cleanup
The claim component generates a link element. To wrap that within a button is not html5 conform, also the link is not really the correct element here. To still have the text label besides the icon, a new prop is used in DpClaim. Also, some flex classes are used to ensure vertical alignment also for the loading state.

Remove "optimistic claiming"
The optimistic behavior (that was implemented before) is not the standard behavior within demosPlan. Also, as we show a loading indicator on the claim button, the rest of the UI should be in sync with that. The alternative would have been to also remove the loading indicator, hence hiding the loading completely. However as we do not use this pattern at all, why should we use it here?