Remove pattern validation for city (#1088)
It does not add value for users to restrict the input for the "city" field based on 
assumptions that can prove wrong anytime. It will do little harm to remove the 
restriction (which will eventually lead to a typo going unnoticed) but it will 
cause massive trouble for users failing for a too-restrictive validation.