fix(refs 754383): hide visibility options if 1 option only
If there is only one role group which should be able to
view faq articles, it makes no sense to force editors
to always chose this option - this one group is activated
silently, and users may control visibility by using the
"Status" toggle.

The permissions used to control which options are shown,
are not sufficient, so a new configuration parameter is
established.