fix(refs 754383): hide visibility options if only one option is allowed (#1109)
If there is only one role group which should be able to view faq articles, it makes 
no sense to force editors to always choose this option. To improve UX, this one 
group is activated silently, and users may control visibility by using the "Status" toggle.

The permission feature_institution_participation which was used to control if one 
of the options are shown, is not sufficient, so instead of using permissions here, 
a new configuration parameter role_groups_faq_visibility is established - because,
technically what is controlled here is not a permission but project configuration.

As the "faq_admin_new" template duplicates most of the "faq_admin_edit" 
template (and because we use the edit view for both edit and create actions
also in other places), this is refactored here as well.