refactor (refs T30254): detect invalid state early
`getFieldMapping` did return a non-usable array format
for specific cases. One call was adjusted so that an
exception is thrown instead of continuing with the
invalid return.

Also move `loadFieldMapping()` call up, so that it is
loaded even if an empty string is passed as parameter
instead of a class string.