fix (refs T22439): Empty internId of original statement in case of last chlid (normal statement) will be deleted.
To "free" the previous given internId and allow to use them again.
Extract deleteStatementObject() into own servcie to avoid making the StatementService even bigger.
Add new permission to decide if the internId of the related originalSTN should be nulled.
This is also the first step do acutally automatically delete the related original statement in case of deleting the last child statement.