fix(refs T30572): use correct button label
The secondary button actually resets the form fields
to the value that was last saved, it discards unsaved changes.
To reflect this, instead of "Abbrechen" the new label
"Änderungen verwerfen" is applied. It has been chosen because
- it explicitly says what the button is doing
- it is already used in similar places within demosPlan
Even if "Verwerfen" is shorter, there is a chance that users
might assume the whole texts to be removed when clicking the button.