feat: remove rollup migration preparation code
Before performing the database rollup in core demosplan we needed
to make sure that the database is in any project in a consistent state.
This is not necessary anymore. In fact, on every migration any project
migrations later than the rollup migration was downmigrated and after the
core migrations reapplied. This commit sets an end to this, as rollup is
performed in most of the existing projects.