Feat(refs: T32012): For now put the import of swagger-ui.css in a comment to evacuate the error
Error:
Build encountered errors
Module Error (from ./node_modules/vue-loader/dist/templateLoader.js):

VueCompilerError: <Transition> expects exactly one child element or component.

Solution:
after the v-if shall follow directly a v-else-if, if it has a relation to the v-if and if it has an argument or modifier.
its important to use them together followed to each other to consider them as a connected block. these are conditional directives.

just in the case for a fallback for any expression v-else can be used