feat: extract initialize logic to event subscriber
Remaining `initialize` in `SessionHandler` did two things:
- determine procedure and add it to the permissions
- check permissions

The first task was removed to `DetermineProcedureSubscriber`, the
permission check was moved from the `SessionHandler` into the
`InitializeService`. This may later be refactored even more.

With this change the `SessionHandler` is only responsible for
storing the Session and for logging the user out.