feat: remove custom DemosToken and use specific SecurityUser
This commit removes the custom DemosToken that was needed to
glue existing authentication to symfony authentication. Now
it is possible to move one step more towards the default
system.
To avoid the regression that again the whole user object is
stored in the session a new SecurityUser is introduced that
holds only information needed for the authentication and is
transformed to the 'real' User later on. This is considered
best practice according to symfony core devs.