Feat(refs:32012): Slot ref in this way is deprecated. Change slot to name to reference to its components slot, use the shortened way of 'v-slot:' to '#'
(Emitted value instead of an instance of Error) (deprecation COMPILER_NATIVE_TEMPLATE) <template> with no special directives
will render as a native template element instead of its inner content in Vue 3.

Here was the problem, that it didnt understand that here its referencing to a slot, because in vue3 slot="" is deprecated.
So it says template tag isnt needed.
if you would like use slot here, instead of slot="" take the name of the slot, to which you would like reference
or take just the component tag, where you would like put content and which is including a slot
it will automatically put the content in the slot of the component even if there is no reference name
use the shortened way of v-slot: to #