fix(refs T32572): fix special characters appearing at the end of string instead at the beginning
The value `rtl` of the `dir` attribute was the reason that special characters appear at the end of the string even if the input was special characters first.

Relating to https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/dir?retiredLocale=de, the auto value should be used for data with an unknown directionality, like data coming from user input, eventually stored in a database. Since this is the case here, the attribute's value has been set accordingly to `auto`.