feat(refs T21517): Adapt DpButton usage to new dpui version (#1436)
demosplan-ui 0.1.3 introduces a new button variant, "subtle".
As the styles for demosplan-ui are currently implememted
in demosplan-core, also the new variant is implemented here.
Since demosplan-ui 0.1.3 drops slots, these places have to be adapted
to use icons instead of slots. All needed icons are already implemented
in demosplan-ui.

When icons come with a proportion signifier class, we
can further tweak spacing of icons inside buttons. this will
eventually be released later on.