feat(refs T32104): tweak map attribution placeholder
This commit changes the way map attributions are handled.
It improves UX by implementing preview of attribution settings and
enhances attributions to support automatic displaying the current year.

1. Default attribution via translation key

The translation key "map.attribution.default" may now contain
{currentYear} and {linkImprint}. This way, projects that require
the link to the imprint can carry on using it this way, while
projects that want to display the current year can do this magically.

Whenever no other attribution is defined, the maps in demosPlan
will now display this attribution.

2. Preview of attribution in Customer Settings

As the other map options (layer, layerUrl) are also reactive,
the attribution is also implemented as a live preview here.

3. Preview of attribution in Procedure Settings

Here, the attribution is only updated on save, because the input
field is further away from the map, hence no direct correlation
must be made.