feat (refs T29585): adds controller_paths for addons
This new (optional) parameter in the addon-manifest allows addons to specify the locations of controllers so that the core can add the routes from annotations to the core router. This opens up the possibility of non-standard namespaces for controllers. Addons can either specifiy one path as a string or multiple paths as an array. If the parameter is not used, it defaults to the Symfony standard of `/src/Controller` which is the behaviour we used before. All addons should still work the same.