refactor: move logic from `EntityFetcher` into resource types where appropriate
This is a preparation for the update to EDT 0.18.0.

In the future different resource type implementations may use
different kinds of databases, requiring different logic to handle
requests. Determining the kind of database outside the resource type
implementation is hard and using an object-oriented approach was
determined to be the better one. Thus, the logic to retrieve
entities was moved into the `DplanResourceType` for now, which
may become a `DoctrineResourceType` in the future or be split
further.