Change config to use tailwind (#1521)
To simplify and modernize our frontend stack, the team
decided to move to Tailwind - it is widely adopted by
the community, configurable, it is well documented,
and it follows our utility-first approach.

To use Tailwind with both demosplan-ui and demosplan-core,
and also be able to configure Tailwind with the demosplan-ui
design tokens, we have chosen the approach of generating
an unpurged Tailwind build within demosplan-ui, which can then
be consumed by demosplan-core, where it would be purged
via postCss.

Purging (and overall processing) of a 12MB css file slowed
down our build drastically. To mitigate that, some things
were needed:

- demosplan-ui.css (which contains Tailwind) gets a dedicated entrypoint
- all 3rd party css is excluded from postCss in dev builds
- the faster sass-embedded is used to compile scss, as fibers
  is not supported on node16+. This needed changing back 
  to `@import '~module'` as sass-embedded runs into problems
  when resolving @import otherwise
- sourcemaps are disabled for scss, however they did not work before

Also, we can use Preflight (the Tailwind normalize) instead of
normalize.css and the inuit-5 base styles.