fix(refs T33234): Remove incorrect late permission evaluation
The addon hook configurations do not yet provide permission information.
If we were to evaluate those, we should filter the scripts in the RPC
method instead of loading all scripts on the users' machine and evaluating
on the client side.

In addition, previously no import tabs would have been loadable from addons
at all if the excel import were disabled.