feat(refs T33234): Add rudimentary support for exposing addon permissions
Addon permissions are not configured in the classic array way the core permissions
are evaluated in. Thus, all addon permissions, while having been marked as exposed,
were dropped when building the frontend permission list.

This change is only a rudimentary solution to exposing the addon permissions as
the correct way would be to redo the permission checking in the frontend to
include addon names and eventually use something more akin to hasPermission(name, context)
with the context being core or an addon name