feat: adds attribute check to permission listener
This includes the new dplanPermisson attribute in permission listener to enable its usage. For the time being, both the annotation as well as the attribute can be used. But already the attribute would take precedent if it exists. This way, we can just delete the old annotation way when it is no longer needed as it acts as a kind of fallback already.

I also removed the old warning, that the annotation should be used as I don't think it is relevant anymore and it made the code much easier. There is also a logging in the permission processing when an empty array is handed in as permissions which would catch the same cases as before anyway.