WIP (refs T30603): works on used classnames
In this form if correctly changes the class names of all classes used somewhere to the correct fqcn, but without using imports (which can be handled by cs-fixer in another step). Sadly this does not affect any use statements or the namespace of the class itself.

I think FullyQualified might be the wrong node at this point. It seems almost correct, but I'm doubtful if it can get all the way there.